src/app/services/kbd-handler-router.service.spec.ts
/// <reference path="../../../typings/index.d.ts" />
/* tslint:disable:no-unused-variable */

import { TestBed, async, inject } from '@angular/core/testing';
import { KbdHandlerRouterService } from './kbd-handler-router.service';
import { CameraKbdHandlerService } from './camera-kbd-handler.service';
import { AsteroidsKbdHandler } from '../inner-games/asteroids/asteroids-kbd-handler';
// import { AsteroidsGame, AsteroidsGameProvider } from '../inner-games/asteroids/asteroids-game';
import { AsteroidsGame } from '../inner-games/asteroids/asteroids-game';
import { BaseService } from './base.service';
import { Ship } from '../inner-games/asteroids/ship';

//TODO: fix this
// getting dreaded 'Error: No provider for jb!' when running theses tests
// so exclude them for now
xdescribe('Service: KbdHandlerRouter', () => {
  beforeEach(() => {
    TestBed.configureTestingModule({
      providers: [
        KbdHandlerRouterService, 
        CameraKbdHandlerService, 
        AsteroidsKbdHandler,
        BaseService,
        // Ship,
        AsteroidsGame
         ]
    });
  });

  it('should instantiate properly', inject(
    [
      KbdHandlerRouterService, 
      CameraKbdHandlerService, 
      AsteroidsKbdHandler, 
      // AsteroidsGameProvider, 
      // BaseService
      ], 
      (service: KbdHandlerRouterService 
      //  cameraKbdHandlerService: CameraKbdHandlerService
       ) => {
    expect(service).toBeTruthy();
    expect(service.cameraKbdHandler).toBeTruthy();
    expect(service.keyEventHandler).toBeTruthy();
    expect(service.toggleKey).toEqual('T'.charCodeAt(0));
  }));

  it('should toggle properly', inject(
    [KbdHandlerRouterService, CameraKbdHandlerService, AsteroidsKbdHandler, BaseService], 
      (service: KbdHandlerRouterService ) => {

        let origKbdHandler = service.activeKbdHandler._name;
        console.log(`origKbdHandler=${origKbdHandler}`);

        // now send a 'T' key event to toggle
        //  service.keyEventHandler(Object.defineProperty(event, 'keyCode', {'value': 'T'.charCodeAt(0)}));
        let evt = <any>{};
        //  evt.keyCode = {};
        //  evt.keyCode = {'value' : 'T'.charCodeAt(0)};
        evt.keyCode = 'T'.charCodeAt(0);
        //  service.keyEventHandler({'keyCode', {'value': 'T'.charCodeAt(0)} });
        service.keyEventHandler(evt);

        let newKbdHandler = service.activeKbdHandler._name;
        console.log(`newKbdHandler=${newKbdHandler}`);

        expect(newKbdHandler).not.toEqual(origKbdHandler);

        // and toggle it back again
        service.keyEventHandler(evt);

        newKbdHandler = service.activeKbdHandler._name;
        console.log(`newKbdHandler=${newKbdHandler}`);

        expect(newKbdHandler).toEqual(origKbdHandler);
  }));
});
